+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Targeted Fix v3
+
+- Add prop format support
+- Improvement to handle the value "" as deleted.
+
 ## Targeted Fix v2
 
 - Update to latest source from PIFork v14